Refactor image_policy_*.py
1. lib/ndfc_python/image_policy.py

- Move all image_policy classes into this shared library

2. Refactor the following files to use shared libraries above

- examples/image_policy_create.py
- examples/image_policy_delete.py
- examples/image_policy_info.py
- examples/image_policy_replace.py
